feat: refactor validation logic and move to shared module
