+++ apps/client/src/shared/index.ts
@@ -1,1 +1,3 @@
 export * from './base.dto';
+export * from './validation';
+export * from './date.util';
