+++ apps/client/src/shared/validation/validation-status.type.ts
@@ -1,0 +1,6 @@
+export type ValidationStatus = 'INITIAL' | 'PENDING' | 'VALID' | 'INVALID';
+
+export interface ValidationStatusWithMessage {
+  status: ValidationStatus;
+  message?: string;
+}
